Makefile:  - fix for install on MacOSX
rest:
 - fixed a gcc compile warning where the cause was that size_t is not defined as unsigned int



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1180 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/squadt/gui_specification_properties.cpp
@@ -132,7 +132,7 @@ namespace squadt {
             try {
               size_t size = boost::filesystem::file_size(full_path);
 
-              outputs->SetItem(c, 3, wxString::Format(wxT("%u"), size));
+              outputs->SetItem(c, 3, wxString::Format(wxT("%u"), static_cast < unsigned int > (size)));
             }
             catch (...) {
               outputs->SetItem(c, 3, wxT("0"));
